init: cron testing
