+++ functions/cronDrawEthereum.ts
@@ -1,8 +1,8 @@
-import { cronAppeal } from "./util/cronAppeal";
+import { cronDraw } from "./util/cronDraw";
 import { Handler, schedule } from "@netlify/functions";
 
 const handler: Handler = async () => {
-    return await cronAppeal("ethereum");
+    return await cronDraw("ethereum");
 };
 
 module.exports.handler = schedule("@hourly", handler);
